feat: make minCropSize configurable through cropperStyle (#66)

+++ library/core/src/commonMain/kotlin/com/attafitamim/krop/core/utils/Rect.kt
@@ -92,7 +92,7 @@ fun IntRect.constrainOffset(bounds: IntRect): IntRect {
 fun Rect.resize(
     handle: Offset,
     delta: Offset,
-    maxZoomSize: Int,
+    maxZoomSize: Float,
 ): Rect {
     var (l, t, r, b) = this
     val (dx, dy) = delta
